working navigation functionality for employee and hr eval pages
