+++ src/components/performance-evaluation/PerfEvalForm.tsx
@@ -1,173 +1,87 @@
-import { useState } from "react";
-import DropdownSelect from "@/components/DropdownSelect";
+"use client"
+import React, { useState} from 'react'
+import { TextAreaWithDescription } from '../Textarea'
+import { PerformanceRatingSlider } from './PerformanceRatingSliders'
 
-interface FormData {
-  id: string; //pulled from logged in user
-  year: number; // year of evaluation
-  a: string; // objective a
-  aComm: string; // comments for objective a
-  b: string; // objective b
-  bComm: string; // comments for objective b
-  c: string; // objective c
-  cComm: string; // comments for objective c
-  d: string; // objective d
-  dComm: string; // comments for objective d
+interface HRServicesProps {
+  username: string;
+  userRole: string;
 }
 
-export default function PerfEvalForm() {
-  // dropdown variables
-  const rating = ["not met", "met", "exceed"];
-  const placeholder = "Select rating";
-  // form variables
-  const [formData, setFormData] = useState<FormData>({
-    id: "00000000-0000-0000-0000-000000000001",
-    year: new Date().getFullYear(),
-    a: "",
-    aComm: "",
-    b: "",
-    bComm: "",
-    c: "",
-    cComm: "",
-    d: "",
-    dComm: "",
-  });
+export default function PerfEvalForm({ username, userRole }: HRServicesProps) {
 
-  const handleChange = (
-    e:
-      | React.ChangeEvent<HTMLSelectElement>
-      | React.ChangeEvent<HTMLInputElement>
-      | React.ChangeEvent<HTMLTextAreaElement>,
-  ) => {
-    const { id, value } = e.target;
-    setFormData({ ...formData, [id]: value });
-  };
+  let employee;
 
-  const onSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    console.log(formData);
-  };
+  if (username === 'Employee') {
+    employee = username
+  }
+
+  const [year, setYear] = useState(userRole === "Employee" ? "2025" : "")
+  const [strengths, setStrengths] = useState("")
+  const [weaknesses, setWeaknesses] = useState("")
+  const [improvements, setImprovements] = useState("")
+  const [notes, setNotes] = useState("")
+  const [ratings, setRatings] = useState<number[]>([])
 
   return (
-    <div className="m-6 rounded-lg border border-light-maroon bg-white p-6 shadow-lg">
-      <form onSubmit={onSubmit}>
-        <div className="mb-4">
-          <label htmlFor="year">Year</label>
-          <input
-            type="number"
-            id="year"
-            defaultValue={formData.year}
-            max="9999"
-            onChange={handleChange}
-            className="block w-24 rounded-lg border border-gray-200 p-2 text-black"
-            required
+    <div>
+     <div className="flex flex-col gap-6 mb-6">
+        <h3>Create Employee Performance Evaluation</h3>
+        <div className="flex gap-2">
+          <input 
+            type="text" 
+            className="border-2 border-gray-400 bg-white text-gray-400 px-3 py-2 rounded-3xl max-w-80" 
+            placeholder={userRole === "Employee" ? username : "Search Employee"}
+            disabled={userRole === "Employee"}
           />
+          <button className="border-2 text-gray-400 px-3 py-2 rounded-3xl hover:text-black hover:border-black transition-all" disabled={userRole === "Employee"}>+</button>
         </div>
-        <div className="mb-4">
-          <label htmlFor="a">
-            A. Employee contributes to the success of Odyssey’s mission and
-            demonstrates core values
-          </label>
-          <DropdownSelect
-            id="a"
-            options={rating}
-            placeholder={placeholder}
-            required={true}
-            onChange={handleChange}
-            className="my-2 rounded-lg border border-gray-200 p-2"
-          />
-          <label htmlFor="aComm">Comments:</label>
-          <textarea
-            id="aComm"
-            placeholder="Add comments..."
-            required={true}
-            rows={4}
-            onChange={handleChange}
-            className="mt-2 block w-full rounded-lg border border-gray-200 p-2 text-black"
-          />
+        <div className="w-36">
+            <h3 className="mb-2">Year</h3>
+            <div className="border-2 text-gray-400 bg-white px-3 py-2 rounded-3xl">
+                <select id="year" name="year" disabled={userRole === "Employee"}>
+                    <option value="" disabled>Select Year</option>
+                    <option value="2020" onChange={() => setYear("2020")}>2020</option>
+                    <option value="2021">2021</option>
+                    <option value="2022">2022</option>
+                    <option value="2023">2023</option>
+                    <option value="2024">2024</option>
+                    <option value="2025" selected={userRole === "Employee"}>2025</option>
+                </select>
+            </div>
         </div>
-        <div className="mb-4">
-          <label htmlFor="b">
-            B. Employee supports and sustains a team approach
-          </label>
-          <DropdownSelect
-            id="b"
-            options={rating}
-            placeholder={placeholder}
-            required={true}
-            onChange={handleChange}
-            className="my-2 rounded-lg border border-gray-200 p-2"
-          />
-          <label htmlFor="bComm">Comments:</label>
-          <textarea
-            id="bComm"
-            placeholder="Add comments..."
-            required={true}
-            rows={4}
-            onChange={handleChange}
-            className="mt-2 block w-full rounded-lg border border-gray-200 p-2 text-black"
-          />
-        </div>
-        <div className="mb-4">
-          <label htmlFor="c">
-            C. Employee cultivates an atmosphere that emphasizes ongoing
-            learning and competence through training, supervision, consultation
-            and feedback
-          </label>
-          <DropdownSelect
-            id="c"
-            options={rating}
-            placeholder={placeholder}
-            required={true}
-            onChange={handleChange}
-            className="my-2 rounded-lg border border-gray-200 p-2"
-          />
-          <label htmlFor="cComm">Comments:</label>
-          <textarea
-            id="cComm"
-            placeholder="Add comments..."
-            required={true}
-            rows={4}
-            onChange={handleChange}
-            className="mt-2 block w-full rounded-lg border border-gray-200 p-2 text-black"
-          />
-        </div>
-        <div className="mb-4">
-          <label htmlFor="d">
-            D. Employee has mastered the responsibilities as outlined in job
-            description. Achievement of Goals (Last 12 Months)
-          </label>
-          <DropdownSelect
-            id="d"
-            options={rating}
-            placeholder={placeholder}
-            required={true}
-            onChange={handleChange}
-            className="my-2 rounded-lg border border-gray-200 p-2"
-          />
-          <label htmlFor="dComm">Comments:</label>
-          <textarea
-            id="dComm"
-            placeholder="Add comments..."
-            required={true}
-            rows={4}
-            onChange={handleChange}
-            className="mt-2 block w-full rounded-lg border border-gray-200 p-2 text-black"
-          />
-        </div>
-        <div className="mb-4">
-          <label htmlFor="signature">E-signature:</label>
-          <input
-            type="text"
-            id="signature"
-            placeholder="Full Name"
-            className="block rounded-lg border border-gray-200 p-2 text-black"
-            required
-          />
-        </div>
-        <button type="submit" className="mt-2 p-2 hover:text-light-maroon">
-          Submit
-        </button>
-      </form>
+      </div>
+
+      <div className="grid grid-cols-1 md:grid-cols-2 gap-4 mb-6">
+        <TextAreaWithDescription
+          label="Strengths"
+          placeholder="List employee's strengths."
+          onChange={(value) => setStrengths(value)}
+        />
+        <TextAreaWithDescription
+          label="Weaknesses"
+          placeholder="List employee's weaknesses."
+          onChange={(value) => setStrengths(value)}
+        />
+        <TextAreaWithDescription
+          label="Improvements"
+          placeholder="List things the employee can improve on."
+          onChange={(value) => setImprovements(value)}
+        />
+        <TextAreaWithDescription
+          label="Other Notes (Optional)"
+          placeholder="Other notes you want to include."
+          onChange={(value) => setNotes(value)}
+        />
+      </div>
+
+      <div className="w-[50%] mb-6">
+        <PerformanceRatingSlider 
+          categories={['Communication', 'Leadership', 'Timeliness', 'Skill', 'Skill', 'Skill']} 
+          onChange={(value: number[]) => setRatings(value)}
+        />
+      </div>
+      <button type="submit" className="border-2 text-gray-400 px-3 py-2 rounded-3xl hover:text-black hover:border-black transition-all">Submit</button>
     </div>
-  );
+  )
 }
