+++ src/components/performance-evaluation/RatingItem.tsx
@@ -9,10 +9,10 @@ export const RatingItem: React.FC<RatingItemProps> = ({ label, percentage }) => 
     return (
       <div>
         <div className="w-32 text-sm text-gray-600">{label}</div>
-        <div className="flex items-center space-x-2 mb-2">
-            <div className="flex-1 bg-gray-200 rounded-full h-2.5">
+        <div className="flex items-center space-x-2 mb-2 space-y-2">
+            <div className="flex-1 bg-gray-200 rounded-full h-5">
             <div 
-                className="bg-gray-400 h-2.5 rounded-full" 
+                className="bg-gray-400 h-5 rounded-full" 
                 style={{ width: `${percentage}%` }}
             ></div>
             </div>
